+++ modules/encrypt-browser/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-preview.2](/compare/@aws-crypto/encrypt-browser@0.1.0-preview.1...@aws-crypto/encrypt-browser@0.1.0-preview.2) (2019-07-24)
+
+
+### Bug Fixes
+
+* browser framed encryption (#156) a2f2ed9, closes #156 #155
+* encrypt/decrypt browser 21d65d0
+* frame length can not be 0 (#149) dc1f92e, closes #149 #129
+
+
+
+
+
 # [0.1.0-preview.1](/compare/@aws-crypto/encrypt-browser@0.1.0-preview.0...@aws-crypto/encrypt-browser@0.1.0-preview.1) (2019-06-21)