+++ modules/integration-node/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-preview.3](/compare/@aws-crypto/integration-node@0.2.0-preview.2...@aws-crypto/integration-node@0.2.0-preview.3) (2019-07-24)
+
+
+### Features
+
+* Encryption tests for integration-node (#153) d7b5e73, closes #153
+
+
+
+
+
 # [0.2.0-preview.2](/compare/@aws-crypto/integration-node@0.2.0-preview.1...@aws-crypto/integration-node@0.2.0-preview.2) (2019-06-22)
 
 **Note:** Version bump only for package @aws-crypto/integration-node