+++ modules/kms-keyring-node/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-preview.2](/compare/@aws-crypto/kms-keyring-node@0.1.0-preview.1...@aws-crypto/kms-keyring-node@0.1.0-preview.2) (2019-07-24)
+
+
+### Bug Fixes
+
+* aws sdk version dependencies pollution (#145) d73d50d, closes #145 #136 #138
+
+
+
+
+
 # [0.1.0-preview.1](/compare/@aws-crypto/kms-keyring-node@0.1.0-preview.0...@aws-crypto/kms-keyring-node@0.1.0-preview.1) (2019-06-21)