+++ modules/material-management-node/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0-preview.2](/compare/@aws-crypto/material-management-node@0.1.0-preview.1...@aws-crypto/material-management-node@0.1.0-preview.2) (2019-07-24)
+
+
+### Features
+
+* Node.js Typescript version dependency (#146) 9dfa857, closes #146 #135 #74
+
+
+
+
+
 # [0.1.0-preview.1](/compare/@aws-crypto/material-management-node@0.1.0-preview.0...@aws-crypto/material-management-node@0.1.0-preview.1) (2019-06-21)